Move the question from Home to the Quiz page
