+++ src/App.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import HomePage from "./components/HomePage";
 import PlaceholderPage from "./components/PlaceholderPage";
 import PageLayout from "./components/PageLayout";
+import Quiz from "./components/Quiz";
 
 import { ApiContext, DefaultApiClient } from "./api";
 import { pathForPage } from "./pages";
@@ -13,7 +14,7 @@ function App() {
         <Routes>
           <Route element={<PageLayout />}>
           <Route path={pathForPage('home')} element={<HomePage/>} />
-          <Route path={pathForPage('quiz')} element={<PlaceholderPage title='Quiz' />} />
+          <Route path={pathForPage('quiz')} element={<Quiz />} />
           <Route path={pathForPage('how_it_works')} element={<PlaceholderPage title='How it works' />} />
           <Route path={pathForPage('leaderboard')} element={<PlaceholderPage title='Leaderboard' />} />
           <Route path={pathForPage('practice')} element={<PlaceholderPage title='Practice' />} />
