+++ src/components/HomePage.tsx
@@ -4,18 +4,9 @@ import { useApiClient } from "../api";
 import Question from "./Question";
 
 const HomePage = () => {
-  const apiClient = useApiClient();
-  const [question, setQuestion] = React.useState<QuestionType>();
 
-  React.useEffect(() => {
-    (async () => {
-      // Load Random Question
-      const question = await apiClient.getRandomQuestion();
-      setQuestion(question);
-    })();
-  }, [apiClient]);
 
-  return <Question question={question} />;
+  return <h1>Welcome to Bnomial!</h1>;
 };
 
 export default HomePage;
